db connections

+++ review-service/src/main/java/com/handson/reviewservice/service/ReviewServiceImpl.java
@@ -6,27 +6,31 @@ import com.handson.reviewservice.persistence.ReviewEntity;
 import com.handson.reviewservice.persistence.ReviewRepository;
 import com.handson.util.exception.InvalidInputException;
 import com.handson.util.exceptionHandler.ServiceUtil;
+import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 
 import java.util.List;
+import java.util.function.Supplier;
+
+import static java.util.logging.Level.FINE;
 
 @RestController
 public class ReviewServiceImpl implements ReviewService {
-
     private static final Logger LOG = LoggerFactory.getLogger(ReviewServiceImpl.class);
-
     private final ReviewRepository repository;
-
     private final ReviewMapper mapper;
-
     private final ServiceUtil serviceUtil;
+    private final Scheduler scheduler;
 
     @Autowired
-    public ReviewServiceImpl(ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+    public ReviewServiceImpl(Scheduler scheduler, ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+        this.scheduler = scheduler;
         this.repository = repository;
         this.mapper = mapper;
         this.serviceUtil = serviceUtil;
@@ -34,35 +38,40 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public Review createReview(Review body) {
+        if (body.getProductId() < 1) throw new InvalidInputException("Invalid productId: " + body.getProductId());
         try {
             ReviewEntity entity = mapper.apiToEntity(body);
             ReviewEntity newEntity = repository.save(entity);
-
             LOG.debug("createReview: created a review entity: {}/{}", body.getProductId(), body.getReviewId());
             return mapper.entityToApi(newEntity);
-
         } catch (DataIntegrityViolationException dive) {
             throw new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Review Id:" + body.getReviewId());
         }
     }
 
     @Override
-    public List<Review> getReviews(int productId) {
-
+    public Flux<Review> getReviews(int productId) {
         if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
-
-        List<ReviewEntity> entityList = repository.findByProductId(productId);
-        List<Review> list = mapper.entityListToApiList(entityList);
-        list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
-
-        LOG.debug("getReviews: response size: {}", list.size());
-
-        return list;
+        LOG.info("Will get reviews for product with id={}", productId);
+        return asyncFlux(() -> Flux.fromIterable(getByProductId(productId))).log(null, FINE);
     }
 
     @Override
     public void deleteReviews(int productId) {
+        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
         LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
         repository.deleteAll(repository.findByProductId(productId));
+    }
+
+    private List<Review> getByProductId(int productId) {
+        List<ReviewEntity> entityList = repository.findByProductId(productId);
+        List<Review> list = mapper.entityListToApiList(entityList);
+        list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
+        LOG.debug("getReviews: response size: {}", list.size());
+        return list;
+    }
+
+    private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+        return Flux.defer(publisherSupplier).subscribeOn(scheduler);
     }
 }
